refactor: rename options.ts to options-execa.ts
Prepares for v0.2's tinyexec adapter which adds a sibling
options-tinyexec.ts. Updates user-facing error messages to
reference the backlog instead of stale v0.1/v0.2 roadmap
claims.
